divide into app (run express) & index (run server)
